requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
openpyxl==3.1.2
SQLAlchemy==2.1.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
